ev: refactor EVUtils (vsp) to ElectricVehicleSpecifications (ev)
