+++ contribs/ev/src/main/java/org/matsim/contrib/ev/fleet/ElectricVehicleSpecifications.java
@@ -1,10 +1,9 @@
-/* *********************************************************************** *
+/*
+ * *********************************************************************** *
  * project: org.matsim.*
- * Controler.java
- *                                                                         *
  * *********************************************************************** *
  *                                                                         *
- * copyright       : (C) 2007 by the members listed in the COPYING,        *
+ * copyright       : (C) 2022 by the members listed in the COPYING,        *
  *                   LICENSE and WARRANTY file.                            *
  * email           : info at matsim dot org                                *
  *                                                                         *
@@ -16,44 +15,26 @@
  *   (at your option) any later version.                                   *
  *   See also COPYING, LICENSE and WARRANTY file                           *
  *                                                                         *
- * *********************************************************************** */
+ * *********************************************************************** *
+ */
 
-package playground.vsp.ev;
+package org.matsim.contrib.ev.fleet;
 
 import java.util.Collection;
 
-import org.matsim.contrib.ev.fleet.ElectricVehicleSpecificationImpl;
 import org.matsim.vehicles.EngineInformation;
 import org.matsim.vehicles.Vehicle;
 
-import com.google.common.collect.ImmutableList;
-
-public class EVUtils {
-	/**
-	 * @param vehicle
-	 * @return the initial energy in kWh
-	 */
-	static Double getInitialEnergy_kWh(Vehicle vehicle) {
-		return (Double)vehicle.getAttributes()
-				.getAttribute(ElectricVehicleSpecificationImpl.INITIAL_ENERGY_kWh);
-	}
-
+public class ElectricVehicleSpecifications {
 	/**
 	 * @param vehicle
 	 * @param initialEnergyInKWh initial energy [kWh]
 	 */
 	public static void setInitialEnergy_kWh(Vehicle vehicle, double initialEnergyInKWh) {
-		vehicle.getAttributes()
-				.putAttribute(ElectricVehicleSpecificationImpl.INITIAL_ENERGY_kWh, initialEnergyInKWh);
-	}
-
-	static ImmutableList<String> getChargerTypes(EngineInformation engineInformation) {
-		return ImmutableList.copyOf((Collection<String>)engineInformation.getAttributes()
-				.getAttribute(ElectricVehicleSpecificationImpl.CHARGER_TYPES));
+		vehicle.getAttributes().putAttribute(ElectricVehicleSpecificationImpl.INITIAL_ENERGY_kWh, initialEnergyInKWh);
 	}
 
 	public static void setChargerTypes(EngineInformation engineInformation, Collection<String> chargerTypes) {
-		engineInformation.getAttributes()
-				.putAttribute(ElectricVehicleSpecificationImpl.CHARGER_TYPES, chargerTypes);
+		engineInformation.getAttributes().putAttribute(ElectricVehicleSpecificationImpl.CHARGER_TYPES, chargerTypes);
 	}
 }
